Refactored smart_switch_controller folder location
Moved smart_switch_controller to the controllers directory
