+++ src/controllers/smart_switch_controller.h
